requirements.txt
langchain
langchain_openai
python-dotenv
fastapi
uvicorn
httpx
openevals
pymupdf
docling
python-multipart